Check if database has been delete

+++ backend/deploy_backend.sh
@@ -7,6 +7,10 @@ log_file=$(date "+%Y_%m_%d-%H_%M_%S")
 
 echo '[-] Removing old database...'
 rm -f db.sqlite3
+if test -e "db.sqlite3"; then
+    echo '[!] Database could not be deleted';
+    exit;
+fi
 
 echo '[-] Applying Django migrations...'
 $python manage.py migrate 2>&1 | tee logs/$log_file.log
